Actually fix ZMQ secret key argument.
We now treat any string argument as the key itself.
Whereas any file-like object will be `.read()` for the key.
This is a breaking change in the API however :/

There's valgrind warnings if the key isn't exactly 40 characters long,
but this is likely also a problem in CSH.

Also small fix to previous commit.